Nice changes
Enrforced uniqueness on usernames in DB
Created public folder for userFiles to get saved in. For this we use keygen to generate a randomized folder name.
Randomized folder name is stored securely in DB
Started implementation of /imgs
Commented out a bunch of no longer needed console.logs
Implimented DB query but it is not working. This will be fixed in a later commit.